Refactor lab files listing
The overall amount of `find` invocations is too big. This leads to
very slow startup on MinGW. The file listing is implemented via single
`find` call and filtering of the resulting list.

Although this implementation works at the speed of the light being
compared with old implementation it has important limitation: only
headers with extensions "h", "hpp" and "hxx" are supported.